pyproject.toml
[project]
name = "postways-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.10",
    "django-environ>=0.12.0",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3, <4.0",
    "redis>=5.2.0,<6.0.0",
    "djangorestframework>=3.15, <3.16",
    "djangorestframework-simplejwt>=5.5.1",
    "channels[daphne]>=4.0,<5.0",
    "django-filter>=25.2",
    "celery>=5.5,<6.0",
]
